--- docs/source/eolearn.ml_tools.postprocessing.rst
@@ -1,7 +1,0 @@
-eolearn.ml\_tools.postprocessing
-================================
-
-.. automodule:: eolearn.ml_tools.postprocessing
-    :members:
-    :undoc-members:
-    :show-inheritance:
