+++ geometry/eolearn/tests/test_morphology.py
@@ -12,11 +12,13 @@ import pytest
 import numpy as np
 from numpy.testing import assert_array_equal
 
-from eolearn.core import FeatureType
-from eolearn.geometry import ErosionTask
+from eolearn.core import EOPatch, FeatureType
+from eolearn.geometry import ErosionTask, MorphologicalFilterTask, MorphologicalOperations, MorphologicalStructFactory
 
 
 CLASSES = (0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10)
+MASK_FEATURE = FeatureType.MASK, "mask"
+MASK_TIMELESS_FEATURE = FeatureType.MASK_TIMELESS, "timeless_mask"
 
 
 @pytest.mark.parametrize("invalid_input", [None, 0, "a"])
@@ -63,3 +65,21 @@ def test_erosion_partial(test_eopatch):
             assert np.sum(mask_after == label) <= np.sum(mask_before == label), "Error in the erosion process"
         else:
             assert_array_equal(mask_after == label, mask_before == label, err_msg="Error in the erosion process")
+
+
+@pytest.mark.parametrize("morph_operation", MorphologicalOperations)
+@pytest.mark.parametrize(
+    "struct_element", [None, MorphologicalStructFactory.get_disk(5), MorphologicalStructFactory.get_rectangle(5, 6)]
+)
+def test_morphological_filter(morph_operation, struct_element):
+    eopatch = EOPatch()
+    eopatch[MASK_FEATURE] = np.random.randint(20, size=(10, 100, 100, 3), dtype=np.uint8)
+    eopatch[MASK_TIMELESS_FEATURE] = np.random.randint(20, 50, size=(100, 100, 5), dtype=np.uint8)
+
+    task = MorphologicalFilterTask(
+        [MASK_FEATURE, MASK_TIMELESS_FEATURE], morph_operation=morph_operation, struct_elem=struct_element
+    )
+    task.execute(eopatch)
+
+    assert eopatch[MASK_FEATURE].shape == (10, 100, 100, 3)
+    assert eopatch[MASK_TIMELESS_FEATURE].shape == (100, 100, 5)
