+++ ml_tools/eolearn/ml_tools/__init__.py
@@ -10,12 +10,6 @@ from .classifier import (
     ImageClassificationMaskTask,
 )
 from .sampling import sample_by_values, BlockSamplingTask, FractionSamplingTask, GridSamplingTask
-from .postprocessing import (
-    MorphologicalOperations,
-    MorphologicalStructFactory,
-    PostprocessingTask,
-    MorphologicalFilterTask,
-)
 from .train_test_split import TrainTestSplitTask
 
 __version__ = "1.0.0"
